# doc/api-doc.yaml
openapi: 3.0.0
info:
  title: API Dokumentation
  description: Dies ist die API Dokumentation für die Anwendung.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Lokaler Server
paths:
...
  /clients:
    post:
      summary: Create a new client
      operationId: createClient
      requestBody:
        description: Request body for creating a new client
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClientRequestBody"
      responses:
        "201":
          description: Client created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: User created
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Invalid request body
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: A user with that email already exists
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Secret encryption failed
components:
  schemas:
...
    CreateClientRequestBody:
      type: object
      properties:
        secret:
          type: string
          example: mysecretpassword
        redirectUrl:
          type: string
          example: http://example.com/callback
      required:
        - secret
        - redirectUrl
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: []